Refactoring run_docker.py and create_folders.py
